pyproject.toml
[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
name = "hemera"
description = "Tools for exporting Ethereum blockchain data to JSON/CSV file and postgresql"
version = "1.0.0a1"
authors = [
    "shanshuo0918 <shuo.shan@lifo.ai>",
    "xuzh <zihao.xu@thehemera.com>",
    "ideal93 <xiang.li@lifo.ai>",
]
readme = "README.md"
license = "Apache-2.0"
classifiers = [
    "Development Status :: 5 - Production/Stable",
    "Intended Audience :: Developers",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
]
keywords = ["ethereum", "indexer", "explorer", "hemera"]
packages = [
    { include = "hemera", from = "." },
    { include = "hemera_udf", from = "." }
]

[tool.poetry.dependencies]
python = ">=3.8,<4"
web3 = "7.8.0"
eth-typing = ">=5.0.0"
eth-utils = "5.1.0"
eth-abi = ">=5.0.1"
python-dateutil = ">=2.8.0,<3"
click = ">=8.0.4,<9"
ethereum-dasm = "0.1.5"
requests = "*"
sqlalchemy = "2.0.31"
psycopg2-binary = "2.9.9"
alembic = "1.13.3"
blinker = ">=1.8.2"
Werkzeug = "3.0.3"
openpyxl = "3.1.3"
redis = ">=4,<6"
urllib3 = ">=2.2.2"
dataclass-wizard = "0.22.3"
pottery = "3.0.0"
orjson = "3.10.7"
mpire = "2.10.2"
dill = "0.3.9"
multiprocess = "0.70.17"
PyYAML = "6.0.2"
fastapi = {version = "<1.0.0,>=0.114.2", extras = ["standard"]}
pydantic = ">=2.4.0"
pydantic-settings = "<3.0.0,>=2.2.1"
sentry-sdk = {version = "<2.0.0,>=1.40.6", extras = ["fastapi"]}
sqlmodel = "*"
pyjwt = "<3.0.0,>=2.8.0"
tenacity = "9.0.0"
kafka-python = "2.0.2"
prometheus_client = "0.21.1"
sortedcontainers = "2.4.0"
deprecated = ">=1.2.13"
pandas = "1.5.3"
numpy = "1.24.4"
packaging = "21.0"

[tool.poetry.group.dev.dependencies]
pytest = ">=7.0.0"
pytest-postgresql = "*"
black = "^21.5b0"
isort = "^5.9.1"
flake8 = "^3.9.2"
ruff = "^0.0.235"
pytest-cov = "*"
pygments = "<2.14.0"

[tool.poetry.scripts]
hemera = "hemera.cli:cli"

[tool.poetry.urls]
"Homepage" = "https://github.com/HemeraProtocol/hemera-indexer"
"Bug Reports" = "https://github.com/HemeraProtocol/hemera-indexer/issues"
"Source" = "https://github.com/HemeraProtocol/hemera-indexer"

[tool.black]
line-length = 120
target-version = ["py38", "py39", "py310", "py311", "py312", "py313"]

[tool.isort]
profile = "black"
multi_line_output = 3
force_single_line = false
include_trailing_comma = true
line_length = 120

[tool.pytest.ini_options]
markers = [
    "api: marks tests as api tests",
    "serial: run this test in single thread",
    "indexer: Tests related to indexer",
    "indexer_exporter: Tests related to the indexer exporter",
    "indexer_utils: Tests related to the indexer utils",
    "indexer_bridge: Tests related to the indexer bridge",
    "indexer_bridge_optimism: Tests related to the indexer bridge optimism",
    "indexer_bridge_arbitrum: Tests related to the indexer bridge arbitrum",
    "indexer_bridge_morph: Tests related to the indexer bridge morph",
    "indexer_jobs: Tests related to the indexer jobs",
    "indexer_jobs_user_ops: Tests related to the indexer jobs user ops",
    "indexer_address_index: Tests related to the indexer address index",
    "explorer_api: Test explorer related API",
    "indexer_jobs_day_mining: Tests related to the indexer day mining",
    "ens: Tests related to the indexer ENS",
    "pipeline: Continuous Integration",
    "multicall_helper: Tests related to the multicall helper",
]
